Changed the user controller and service to match the user guidelines

+++ studdit/apps/studdit/src/app/users/users.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { User } from './schemas/user.schema';
 import { Model } from 'mongoose';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { DeleteUserDto } from './dto/delete-user.dto';
 
 @Injectable()
 export class UsersService {
@@ -35,12 +36,19 @@ export class UsersService {
         }
     }
 
-    async delete(id: string): Promise<User> {
+    async delete(id: string, deleteUserDTO: DeleteUserDto): Promise<User> {
         try {
-            const deletedUser = await this.userModel.findOneAndDelete({ _id: id });
-            if (!deletedUser) {
+            const user = await this.userModel.findOne({ _id: id });
+
+            if (!user) {
                 throw new NotFoundException('User not found');
             }
+
+            if (deleteUserDTO.password !== user.password) {
+                throw new UnauthorizedException('Password is incorrect');
+            }
+
+            const deletedUser = await this.userModel.findOneAndDelete({ _id: id });
             return deletedUser;
         } catch (error) {
             throw new Error('Unable to delete user');
@@ -49,11 +57,18 @@ export class UsersService {
 
     async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
         try {
-            const user = await this.userModel.findOneAndUpdate({ _id: id }, updateUserDto, { new: true });
-
+            const user = await this.userModel.findOne({ _id: id });
+            
             if (!user) {
                 throw new NotFoundException('User not found');
             }
+
+            if (updateUserDto.currentPassword !== user.password) {
+                throw new UnauthorizedException('Current password is incorrect');
+            }
+
+            user.password = updateUserDto.newPassword;
+            await user.save();
 
             return user;
         } catch (error) {
